Move config/template types into library
