+++ xtask/Cargo.toml
@@ -9,5 +9,6 @@ anyhow       = "1.0.93"
 clap         = { version = "4.5.21", features = ["derive"] }
 env_logger   = "0.11.5"
 esp-metadata = { version = "0.4.0", features = ["clap"] }
+esp-generate = { path = ".." }
 log          = "0.4.22"
 tempfile     = "3.14.0"
